requirements.txt
langchain
langchain-openai
langchain-community
streamlit
python-dotenv
langserve
uvicorn
fastapi
sse_starlette
beautifulsoup4
ipykernel
pypdf
chromadb
faiss-cpu
wikipedia
arxiv
groq
langchain-groq
pandas 
numpy 
scikit-learn 
lime 
openpyxl 
matplotlib 
seaborn 
altair 
plotly
sentence-transformers 
huggingface-hub
ChatGroq